Show expanded CRS arguments

Wrappers for sp::CRS and inla.CRS objects to handle the coordinate reference system argument string. These methods should no longer be used with PROJ6/rgdal3; see fm_crs_get_wkt() for a new approach.

fm_CRS_as_list(x, ...)

fm_list_as_CRS(x, ...)

fm_CRSargs(x, ...)

fm_list_as_CRSargs(x, ...)

fm_CRSargs_as_list(x, ...)

Arguments

x
An sp::CRS or inla.CRS object (for fm_CRSargs and fm_CRS_as_list), a character string (for fm_CRSargs_as_list), or a list (for fm_list_as_CRS and fm_list_as_CRSargs).

...
Additional arguments passed on to other methods.

Value

For fm_CRSargs and fm_list_as_CRSargs, a character string with PROJ.4 arguments.

For fm_CRS_as_list and fm_CRSargs_as_list, a list of name/value pairs.

For fm_list_as_CRS, a CRS or inla.CRS object.

For fm_list_as_CRSargs(), a CRS proj4 string for name=value pair list

For fm_CRSargs_as_list(), a list of name=value pairs from CRS proj4 string

See also

Author

Finn Lindgren finn.lindgren@gmail.com

Examples

if (require(rgdal)) {
  crs0 <- fm_CRS("longlat")
  p4s <- fm_CRSargs(crs0)
  lst <- fm_CRSargs_as_list(p4s)
  crs1 <- fm_list_as_CRS(lst)
  lst$a <- 2
  crs2 <- fm_CRS(p4s, args = lst)
  print(fm_CRSargs(crs0))
  print(fm_CRSargs(crs1))
  print(fm_CRSargs(crs2))
}

#> Warning: Use of old predefined projection 'longlat' is deprecated. Converting to 'longlat_norm'

#> Warning: Discarded ellps unknown in CRS definition: +proj=longlat +R=1 +no_defs +type=crs

#> Warning: Discarded datum unknown in CRS definition

#> Warning: 'fm_CRSargs' should not be used with PROJ6 and rgdal v3
#> Call stack for developer debugging:
#> pkgdown::deploy_to_branch(new_process = FALSE)
#> c("build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE, ", "    ...)")
#> c("build_site_local(pkg = pkg, examples = examples, run_dont_run = run_do", "    seed = seed, lazy = lazy, override = override, preview = preview, ", "    devel = devel)")
#> c("build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = ", "    seed = seed, override = override, preview = FALSE, devel = devel)")
#> c("purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, ", "    examples = examples, run_dont_run = run_dont_run)")
#> .f(.x[[i]], ...)
#> c("withCallingHandlers(data_reference_topic(topic, pkg, examples = exampl", "    run_dont_run = run_dont_run), error = function(err) {", "    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), ", "        i = err$message)", "    abort(msg, parent = err)", "})")
#> data_reference_topic(topic, pkg, examples = examples, run_dont_run = r
#> c("run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv(", "    topic = tools::file_path_sans_ext(topic$file_in), run_examples = e", "    run_dont_run = run_dont_run)")
#> highlight_examples(code, topic, env = env)
#> c("downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, ", "    env = child_env(env))")
#> evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> c("evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, ", "    debug = debug, last = i == length(out), use_try = stop_on_error !=", "        2L, keep_warning = keep_warning, keep_message = keep_message, ", "    output_handler = output_handler, include_timing = include_timing)")
#> c("timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, ", "    envir, enclos)), warning = wHandler, error = eHandler, message = m")
#> c("handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, ", "    enclos)), warning = wHandler, error = eHandler, message = mHandler")
#> try(f, silent = TRUE)
#> c("tryCatch(expr, error = function(e) {", "    call <- conditionCall(e)", "    if (!is.null(call)) {", "        if (identical(call[[1L]], quote(doTryCatch))) ", "            call <- sys.call(-4L)", "        dcall <- deparse(call)[1L]", "        prefix <- paste(\"Error in\", dcall, \": \")", "        LONG <- 75L", "        sm <- strsplit(conditionMessage(e), \"\\n\")[[1L]]", "        w <- 14L + nchar(dcall, type = \"w\") + nchar(sm[1L], type = \"w\"", "        if (is.na(w)) ", "            w <- 14L + nchar(dcall, type = \"b\") + nchar(sm[1L], ", 
#> "                type = \"b\")", "        if (w > LONG) ", "            prefix <- paste0(prefix, \"\\n  \")", "    }", "    else prefix <- \"Error : \"", "    msg <- paste0(prefix, conditionMessage(e), \"\\n\")", "    .Internal(seterrmessage(msg[1L]))", "    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {", "        cat(msg, file = outFile)", "        .Internal(printDeferredWarnings())", "    }", "    invisible(structure(msg, class = \"try-error\", condition = e))", "})")
#> tryCatchList(expr, classes, parentenv, handlers)
#> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> doTryCatch(return(expr), name, parentenv, handler)
#> c("withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = ", "    error = eHandler, message = mHandler)")
#> withVisible(eval(expr, envir, enclos))
#> eval(expr, envir, enclos)
#> eval(expr, envir, enclos)
#> fm_CRSargs(crs0)
#> fm_not_for_PROJ6("fm_CRSargs")

#> Warning: Discarded ellps unknown in CRS definition: +proj=longlat +R=1 +no_defs +type=crs

#> Warning: Discarded datum unknown in CRS definition

#> Warning: 'fm_CRSargs_as_list' should not be used with PROJ6 and rgdal v3
#> Call stack for developer debugging:
#> pkgdown::deploy_to_branch(new_process = FALSE)
#> c("build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE, ", "    ...)")
#> c("build_site_local(pkg = pkg, examples = examples, run_dont_run = run_do", "    seed = seed, lazy = lazy, override = override, preview = preview, ", "    devel = devel)")
#> c("build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = ", "    seed = seed, override = override, preview = FALSE, devel = devel)")
#> c("purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, ", "    examples = examples, run_dont_run = run_dont_run)")
#> .f(.x[[i]], ...)
#> c("withCallingHandlers(data_reference_topic(topic, pkg, examples = exampl", "    run_dont_run = run_dont_run), error = function(err) {", "    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), ", "        i = err$message)", "    abort(msg, parent = err)", "})")
#> data_reference_topic(topic, pkg, examples = examples, run_dont_run = r
#> c("run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv(", "    topic = tools::file_path_sans_ext(topic$file_in), run_examples = e", "    run_dont_run = run_dont_run)")
#> highlight_examples(code, topic, env = env)
#> c("downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, ", "    env = child_env(env))")
#> evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> c("evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, ", "    debug = debug, last = i == length(out), use_try = stop_on_error !=", "        2L, keep_warning = keep_warning, keep_message = keep_message, ", "    output_handler = output_handler, include_timing = include_timing)")
#> c("timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, ", "    envir, enclos)), warning = wHandler, error = eHandler, message = m")
#> c("handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, ", "    enclos)), warning = wHandler, error = eHandler, message = mHandler")
#> try(f, silent = TRUE)
#> c("tryCatch(expr, error = function(e) {", "    call <- conditionCall(e)", "    if (!is.null(call)) {", "        if (identical(call[[1L]], quote(doTryCatch))) ", "            call <- sys.call(-4L)", "        dcall <- deparse(call)[1L]", "        prefix <- paste(\"Error in\", dcall, \": \")", "        LONG <- 75L", "        sm <- strsplit(conditionMessage(e), \"\\n\")[[1L]]", "        w <- 14L + nchar(dcall, type = \"w\") + nchar(sm[1L], type = \"w\"", "        if (is.na(w)) ", "            w <- 14L + nchar(dcall, type = \"b\") + nchar(sm[1L], ", 
#> "                type = \"b\")", "        if (w > LONG) ", "            prefix <- paste0(prefix, \"\\n  \")", "    }", "    else prefix <- \"Error : \"", "    msg <- paste0(prefix, conditionMessage(e), \"\\n\")", "    .Internal(seterrmessage(msg[1L]))", "    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {", "        cat(msg, file = outFile)", "        .Internal(printDeferredWarnings())", "    }", "    invisible(structure(msg, class = \"try-error\", condition = e))", "})")
#> tryCatchList(expr, classes, parentenv, handlers)
#> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> doTryCatch(return(expr), name, parentenv, handler)
#> c("withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = ", "    error = eHandler, message = mHandler)")
#> withVisible(eval(expr, envir, enclos))
#> eval(expr, envir, enclos)
#> eval(expr, envir, enclos)
#> fm_CRSargs_as_list(p4s)
#> fm_not_for_PROJ6("fm_CRSargs_as_list")

#> Warning: 'fm_list_as_CRS' should not be used with PROJ6 and rgdal v3
#> Call stack for developer debugging:
#> pkgdown::deploy_to_branch(new_process = FALSE)
#> c("build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE, ", "    ...)")
#> c("build_site_local(pkg = pkg, examples = examples, run_dont_run = run_do", "    seed = seed, lazy = lazy, override = override, preview = preview, ", "    devel = devel)")
#> c("build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = ", "    seed = seed, override = override, preview = FALSE, devel = devel)")
#> c("purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, ", "    examples = examples, run_dont_run = run_dont_run)")
#> .f(.x[[i]], ...)
#> c("withCallingHandlers(data_reference_topic(topic, pkg, examples = exampl", "    run_dont_run = run_dont_run), error = function(err) {", "    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), ", "        i = err$message)", "    abort(msg, parent = err)", "})")
#> data_reference_topic(topic, pkg, examples = examples, run_dont_run = r
#> c("run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv(", "    topic = tools::file_path_sans_ext(topic$file_in), run_examples = e", "    run_dont_run = run_dont_run)")
#> highlight_examples(code, topic, env = env)
#> c("downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, ", "    env = child_env(env))")
#> evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> c("evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, ", "    debug = debug, last = i == length(out), use_try = stop_on_error !=", "        2L, keep_warning = keep_warning, keep_message = keep_message, ", "    output_handler = output_handler, include_timing = include_timing)")
#> c("timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, ", "    envir, enclos)), warning = wHandler, error = eHandler, message = m")
#> c("handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, ", "    enclos)), warning = wHandler, error = eHandler, message = mHandler")
#> try(f, silent = TRUE)
#> c("tryCatch(expr, error = function(e) {", "    call <- conditionCall(e)", "    if (!is.null(call)) {", "        if (identical(call[[1L]], quote(doTryCatch))) ", "            call <- sys.call(-4L)", "        dcall <- deparse(call)[1L]", "        prefix <- paste(\"Error in\", dcall, \": \")", "        LONG <- 75L", "        sm <- strsplit(conditionMessage(e), \"\\n\")[[1L]]", "        w <- 14L + nchar(dcall, type = \"w\") + nchar(sm[1L], type = \"w\"", "        if (is.na(w)) ", "            w <- 14L + nchar(dcall, type = \"b\") + nchar(sm[1L], ", 
#> "                type = \"b\")", "        if (w > LONG) ", "            prefix <- paste0(prefix, \"\\n  \")", "    }", "    else prefix <- \"Error : \"", "    msg <- paste0(prefix, conditionMessage(e), \"\\n\")", "    .Internal(seterrmessage(msg[1L]))", "    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {", "        cat(msg, file = outFile)", "        .Internal(printDeferredWarnings())", "    }", "    invisible(structure(msg, class = \"try-error\", condition = e))", "})")
#> tryCatchList(expr, classes, parentenv, handlers)
#> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> doTryCatch(return(expr), name, parentenv, handler)
#> c("withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = ", "    error = eHandler, message = mHandler)")
#> withVisible(eval(expr, envir, enclos))
#> eval(expr, envir, enclos)
#> eval(expr, envir, enclos)
#> fm_list_as_CRS(lst)
#> fm_not_for_PROJ6("fm_list_as_CRS")

#> Warning: 'fm_CRS' should be given a SRS_string for PROJ6 or a known keyword for a predefined string given in projargs. Using fallback PROJ4 method.

#> Warning: Discarded ellps unknown in CRS definition: +proj=longlat +R=1 +no_defs +type=crs

#> Warning: Discarded datum unknown in CRS definition

#> Warning: 'fm_CRS_as_list' should not be used with PROJ6 and rgdal v3
#> Call stack for developer debugging:
#> pkgdown::deploy_to_branch(new_process = FALSE)
#> c("build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE, ", "    ...)")
#> c("build_site_local(pkg = pkg, examples = examples, run_dont_run = run_do", "    seed = seed, lazy = lazy, override = override, preview = preview, ", "    devel = devel)")
#> c("build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = ", "    seed = seed, override = override, preview = FALSE, devel = devel)")
#> c("purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, ", "    examples = examples, run_dont_run = run_dont_run)")
#> .f(.x[[i]], ...)
#> c("withCallingHandlers(data_reference_topic(topic, pkg, examples = exampl", "    run_dont_run = run_dont_run), error = function(err) {", "    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), ", "        i = err$message)", "    abort(msg, parent = err)", "})")
#> data_reference_topic(topic, pkg, examples = examples, run_dont_run = r
#> c("run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv(", "    topic = tools::file_path_sans_ext(topic$file_in), run_examples = e", "    run_dont_run = run_dont_run)")
#> highlight_examples(code, topic, env = env)
#> c("downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, ", "    env = child_env(env))")
#> evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> c("evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, ", "    debug = debug, last = i == length(out), use_try = stop_on_error !=", "        2L, keep_warning = keep_warning, keep_message = keep_message, ", "    output_handler = output_handler, include_timing = include_timing)")
#> c("timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, ", "    envir, enclos)), warning = wHandler, error = eHandler, message = m")
#> c("handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, ", "    enclos)), warning = wHandler, error = eHandler, message = mHandler")
#> try(f, silent = TRUE)
#> c("tryCatch(expr, error = function(e) {", "    call <- conditionCall(e)", "    if (!is.null(call)) {", "        if (identical(call[[1L]], quote(doTryCatch))) ", "            call <- sys.call(-4L)", "        dcall <- deparse(call)[1L]", "        prefix <- paste(\"Error in\", dcall, \": \")", "        LONG <- 75L", "        sm <- strsplit(conditionMessage(e), \"\\n\")[[1L]]", "        w <- 14L + nchar(dcall, type = \"w\") + nchar(sm[1L], type = \"w\"", "        if (is.na(w)) ", "            w <- 14L + nchar(dcall, type = \"b\") + nchar(sm[1L], ", 
#> "                type = \"b\")", "        if (w > LONG) ", "            prefix <- paste0(prefix, \"\\n  \")", "    }", "    else prefix <- \"Error : \"", "    msg <- paste0(prefix, conditionMessage(e), \"\\n\")", "    .Internal(seterrmessage(msg[1L]))", "    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {", "        cat(msg, file = outFile)", "        .Internal(printDeferredWarnings())", "    }", "    invisible(structure(msg, class = \"try-error\", condition = e))", "})")
#> tryCatchList(expr, classes, parentenv, handlers)
#> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> doTryCatch(return(expr), name, parentenv, handler)
#> c("withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = ", "    error = eHandler, message = mHandler)")
#> withVisible(eval(expr, envir, enclos))
#> eval(expr, envir, enclos)
#> eval(expr, envir, enclos)
#> fm_CRS(p4s, args = lst)
#> fm_CRS_as_list(x)
#> fm_not_for_PROJ6("fm_CRS_as_list")

#> Warning: 'fm_CRSargs_as_list' should not be used with PROJ6 and rgdal v3
#> Call stack for developer debugging:
#> pkgdown::deploy_to_branch(new_process = FALSE)
#> c("build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE, ", "    ...)")
#> c("build_site_local(pkg = pkg, examples = examples, run_dont_run = run_do", "    seed = seed, lazy = lazy, override = override, preview = preview, ", "    devel = devel)")
#> c("build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = ", "    seed = seed, override = override, preview = FALSE, devel = devel)")
#> c("purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, ", "    examples = examples, run_dont_run = run_dont_run)")
#> .f(.x[[i]], ...)
#> c("withCallingHandlers(data_reference_topic(topic, pkg, examples = exampl", "    run_dont_run = run_dont_run), error = function(err) {", "    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), ", "        i = err$message)", "    abort(msg, parent = err)", "})")
#> data_reference_topic(topic, pkg, examples = examples, run_dont_run = r
#> c("run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv(", "    topic = tools::file_path_sans_ext(topic$file_in), run_examples = e", "    run_dont_run = run_dont_run)")
#> highlight_examples(code, topic, env = env)
#> c("downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, ", "    env = child_env(env))")
#> evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> c("evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, ", "    debug = debug, last = i == length(out), use_try = stop_on_error !=", "        2L, keep_warning = keep_warning, keep_message = keep_message, ", "    output_handler = output_handler, include_timing = include_timing)")
#> c("timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, ", "    envir, enclos)), warning = wHandler, error = eHandler, message = m")
#> c("handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, ", "    enclos)), warning = wHandler, error = eHandler, message = mHandler")
#> try(f, silent = TRUE)
#> c("tryCatch(expr, error = function(e) {", "    call <- conditionCall(e)", "    if (!is.null(call)) {", "        if (identical(call[[1L]], quote(doTryCatch))) ", "            call <- sys.call(-4L)", "        dcall <- deparse(call)[1L]", "        prefix <- paste(\"Error in\", dcall, \": \")", "        LONG <- 75L", "        sm <- strsplit(conditionMessage(e), \"\\n\")[[1L]]", "        w <- 14L + nchar(dcall, type = \"w\") + nchar(sm[1L], type = \"w\"", "        if (is.na(w)) ", "            w <- 14L + nchar(dcall, type = \"b\") + nchar(sm[1L], ", 
#> "                type = \"b\")", "        if (w > LONG) ", "            prefix <- paste0(prefix, \"\\n  \")", "    }", "    else prefix <- \"Error : \"", "    msg <- paste0(prefix, conditionMessage(e), \"\\n\")", "    .Internal(seterrmessage(msg[1L]))", "    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {", "        cat(msg, file = outFile)", "        .Internal(printDeferredWarnings())", "    }", "    invisible(structure(msg, class = \"try-error\", condition = e))", "})")
#> tryCatchList(expr, classes, parentenv, handlers)
#> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> doTryCatch(return(expr), name, parentenv, handler)
#> c("withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = ", "    error = eHandler, message = mHandler)")
#> withVisible(eval(expr, envir, enclos))
#> eval(expr, envir, enclos)
#> eval(expr, envir, enclos)
#> fm_CRS(p4s, args = lst)
#> fm_CRS_as_list(x)
#> fm_CRSargs_as_list(fm_CRSargs(x))
#> fm_not_for_PROJ6("fm_CRSargs_as_list")

#> Warning: 'fm_CRSargs' should not be used with PROJ6 and rgdal v3
#> Call stack for developer debugging:
#> pkgdown::deploy_to_branch(new_process = FALSE)
#> c("build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE, ", "    ...)")
#> c("build_site_local(pkg = pkg, examples = examples, run_dont_run = run_do", "    seed = seed, lazy = lazy, override = override, preview = preview, ", "    devel = devel)")
#> c("build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = ", "    seed = seed, override = override, preview = FALSE, devel = devel)")
#> c("purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, ", "    examples = examples, run_dont_run = run_dont_run)")
#> .f(.x[[i]], ...)
#> c("withCallingHandlers(data_reference_topic(topic, pkg, examples = exampl", "    run_dont_run = run_dont_run), error = function(err) {", "    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), ", "        i = err$message)", "    abort(msg, parent = err)", "})")
#> data_reference_topic(topic, pkg, examples = examples, run_dont_run = r
#> c("run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv(", "    topic = tools::file_path_sans_ext(topic$file_in), run_examples = e", "    run_dont_run = run_dont_run)")
#> highlight_examples(code, topic, env = env)
#> c("downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, ", "    env = child_env(env))")
#> evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> c("evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, ", "    debug = debug, last = i == length(out), use_try = stop_on_error !=", "        2L, keep_warning = keep_warning, keep_message = keep_message, ", "    output_handler = output_handler, include_timing = include_timing)")
#> c("timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, ", "    envir, enclos)), warning = wHandler, error = eHandler, message = m")
#> c("handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, ", "    enclos)), warning = wHandler, error = eHandler, message = mHandler")
#> try(f, silent = TRUE)
#> c("tryCatch(expr, error = function(e) {", "    call <- conditionCall(e)", "    if (!is.null(call)) {", "        if (identical(call[[1L]], quote(doTryCatch))) ", "            call <- sys.call(-4L)", "        dcall <- deparse(call)[1L]", "        prefix <- paste(\"Error in\", dcall, \": \")", "        LONG <- 75L", "        sm <- strsplit(conditionMessage(e), \"\\n\")[[1L]]", "        w <- 14L + nchar(dcall, type = \"w\") + nchar(sm[1L], type = \"w\"", "        if (is.na(w)) ", "            w <- 14L + nchar(dcall, type = \"b\") + nchar(sm[1L], ", 
#> "                type = \"b\")", "        if (w > LONG) ", "            prefix <- paste0(prefix, \"\\n  \")", "    }", "    else prefix <- \"Error : \"", "    msg <- paste0(prefix, conditionMessage(e), \"\\n\")", "    .Internal(seterrmessage(msg[1L]))", "    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {", "        cat(msg, file = outFile)", "        .Internal(printDeferredWarnings())", "    }", "    invisible(structure(msg, class = \"try-error\", condition = e))", "})")
#> tryCatchList(expr, classes, parentenv, handlers)
#> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> doTryCatch(return(expr), name, parentenv, handler)
#> c("withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = ", "    error = eHandler, message = mHandler)")
#> withVisible(eval(expr, envir, enclos))
#> eval(expr, envir, enclos)
#> eval(expr, envir, enclos)
#> fm_CRS(p4s, args = lst)
#> fm_CRS_as_list(x)
#> fm_CRSargs_as_list(fm_CRSargs(x))
#> fm_CRSargs(x)
#> fm_not_for_PROJ6("fm_CRSargs")

#> Warning: Discarded ellps unknown in CRS definition: +proj=longlat +R=1 +no_defs +type=crs

#> Warning: Discarded datum unknown in CRS definition

#> Warning: Discarded ellps unknown in CRS definition: +proj=longlat +R=1 +no_defs +type=crs

#> Warning: Discarded datum unknown in CRS definition

#> Warning: 'fm_CRSargs' should not be used with PROJ6 and rgdal v3
#> Call stack for developer debugging:
#> pkgdown::deploy_to_branch(new_process = FALSE)
#> c("build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE, ", "    ...)")
#> c("build_site_local(pkg = pkg, examples = examples, run_dont_run = run_do", "    seed = seed, lazy = lazy, override = override, preview = preview, ", "    devel = devel)")
#> c("build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = ", "    seed = seed, override = override, preview = FALSE, devel = devel)")
#> c("purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, ", "    examples = examples, run_dont_run = run_dont_run)")
#> .f(.x[[i]], ...)
#> c("withCallingHandlers(data_reference_topic(topic, pkg, examples = exampl", "    run_dont_run = run_dont_run), error = function(err) {", "    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), ", "        i = err$message)", "    abort(msg, parent = err)", "})")
#> data_reference_topic(topic, pkg, examples = examples, run_dont_run = r
#> c("run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv(", "    topic = tools::file_path_sans_ext(topic$file_in), run_examples = e", "    run_dont_run = run_dont_run)")
#> highlight_examples(code, topic, env = env)
#> c("downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, ", "    env = child_env(env))")
#> evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> c("evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, ", "    debug = debug, last = i == length(out), use_try = stop_on_error !=", "        2L, keep_warning = keep_warning, keep_message = keep_message, ", "    output_handler = output_handler, include_timing = include_timing)")
#> c("timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, ", "    envir, enclos)), warning = wHandler, error = eHandler, message = m")
#> c("handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, ", "    enclos)), warning = wHandler, error = eHandler, message = mHandler")
#> try(f, silent = TRUE)
#> c("tryCatch(expr, error = function(e) {", "    call <- conditionCall(e)", "    if (!is.null(call)) {", "        if (identical(call[[1L]], quote(doTryCatch))) ", "            call <- sys.call(-4L)", "        dcall <- deparse(call)[1L]", "        prefix <- paste(\"Error in\", dcall, \": \")", "        LONG <- 75L", "        sm <- strsplit(conditionMessage(e), \"\\n\")[[1L]]", "        w <- 14L + nchar(dcall, type = \"w\") + nchar(sm[1L], type = \"w\"", "        if (is.na(w)) ", "            w <- 14L + nchar(dcall, type = \"b\") + nchar(sm[1L], ", 
#> "                type = \"b\")", "        if (w > LONG) ", "            prefix <- paste0(prefix, \"\\n  \")", "    }", "    else prefix <- \"Error : \"", "    msg <- paste0(prefix, conditionMessage(e), \"\\n\")", "    .Internal(seterrmessage(msg[1L]))", "    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {", "        cat(msg, file = outFile)", "        .Internal(printDeferredWarnings())", "    }", "    invisible(structure(msg, class = \"try-error\", condition = e))", "})")
#> tryCatchList(expr, classes, parentenv, handlers)
#> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> doTryCatch(return(expr), name, parentenv, handler)
#> c("withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = ", "    error = eHandler, message = mHandler)")
#> withVisible(eval(expr, envir, enclos))
#> eval(expr, envir, enclos)
#> eval(expr, envir, enclos)
#> print(fm_CRSargs(crs0))
#> fm_CRSargs(crs0)
#> fm_not_for_PROJ6("fm_CRSargs")

#> Warning: Discarded ellps unknown in CRS definition: +proj=longlat +R=1 +no_defs +type=crs

#> Warning: Discarded datum unknown in CRS definition

#> [1] "+proj=longlat +R=1 +no_defs"

#> Warning: 'fm_CRSargs' should not be used with PROJ6 and rgdal v3
#> Call stack for developer debugging:
#> pkgdown::deploy_to_branch(new_process = FALSE)
#> c("build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE, ", "    ...)")
#> c("build_site_local(pkg = pkg, examples = examples, run_dont_run = run_do", "    seed = seed, lazy = lazy, override = override, preview = preview, ", "    devel = devel)")
#> c("build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = ", "    seed = seed, override = override, preview = FALSE, devel = devel)")
#> c("purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, ", "    examples = examples, run_dont_run = run_dont_run)")
#> .f(.x[[i]], ...)
#> c("withCallingHandlers(data_reference_topic(topic, pkg, examples = exampl", "    run_dont_run = run_dont_run), error = function(err) {", "    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), ", "        i = err$message)", "    abort(msg, parent = err)", "})")
#> data_reference_topic(topic, pkg, examples = examples, run_dont_run = r
#> c("run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv(", "    topic = tools::file_path_sans_ext(topic$file_in), run_examples = e", "    run_dont_run = run_dont_run)")
#> highlight_examples(code, topic, env = env)
#> c("downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, ", "    env = child_env(env))")
#> evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> c("evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, ", "    debug = debug, last = i == length(out), use_try = stop_on_error !=", "        2L, keep_warning = keep_warning, keep_message = keep_message, ", "    output_handler = output_handler, include_timing = include_timing)")
#> c("timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, ", "    envir, enclos)), warning = wHandler, error = eHandler, message = m")
#> c("handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, ", "    enclos)), warning = wHandler, error = eHandler, message = mHandler")
#> try(f, silent = TRUE)
#> c("tryCatch(expr, error = function(e) {", "    call <- conditionCall(e)", "    if (!is.null(call)) {", "        if (identical(call[[1L]], quote(doTryCatch))) ", "            call <- sys.call(-4L)", "        dcall <- deparse(call)[1L]", "        prefix <- paste(\"Error in\", dcall, \": \")", "        LONG <- 75L", "        sm <- strsplit(conditionMessage(e), \"\\n\")[[1L]]", "        w <- 14L + nchar(dcall, type = \"w\") + nchar(sm[1L], type = \"w\"", "        if (is.na(w)) ", "            w <- 14L + nchar(dcall, type = \"b\") + nchar(sm[1L], ", 
#> "                type = \"b\")", "        if (w > LONG) ", "            prefix <- paste0(prefix, \"\\n  \")", "    }", "    else prefix <- \"Error : \"", "    msg <- paste0(prefix, conditionMessage(e), \"\\n\")", "    .Internal(seterrmessage(msg[1L]))", "    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {", "        cat(msg, file = outFile)", "        .Internal(printDeferredWarnings())", "    }", "    invisible(structure(msg, class = \"try-error\", condition = e))", "})")
#> tryCatchList(expr, classes, parentenv, handlers)
#> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> doTryCatch(return(expr), name, parentenv, handler)
#> c("withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = ", "    error = eHandler, message = mHandler)")
#> withVisible(eval(expr, envir, enclos))
#> eval(expr, envir, enclos)
#> eval(expr, envir, enclos)
#> print(fm_CRSargs(crs1))
#> fm_CRSargs(crs1)
#> fm_not_for_PROJ6("fm_CRSargs")

#> [1] NA

#> Warning: 'fm_CRSargs' should not be used with PROJ6 and rgdal v3
#> Call stack for developer debugging:
#> pkgdown::deploy_to_branch(new_process = FALSE)
#> c("build_site(pkg, devel = FALSE, preview = FALSE, install = FALSE, ", "    ...)")
#> c("build_site_local(pkg = pkg, examples = examples, run_dont_run = run_do", "    seed = seed, lazy = lazy, override = override, preview = preview, ", "    devel = devel)")
#> c("build_reference(pkg, lazy = lazy, examples = examples, run_dont_run = ", "    seed = seed, override = override, preview = FALSE, devel = devel)")
#> c("purrr::map(topics, build_reference_topic, pkg = pkg, lazy = lazy, ", "    examples = examples, run_dont_run = run_dont_run)")
#> .f(.x[[i]], ...)
#> c("withCallingHandlers(data_reference_topic(topic, pkg, examples = exampl", "    run_dont_run = run_dont_run), error = function(err) {", "    msg <- c(paste0(\"Failed to parse Rd in \", topic$file_in), ", "        i = err$message)", "    abort(msg, parent = err)", "})")
#> data_reference_topic(topic, pkg, examples = examples, run_dont_run = r
#> c("run_examples(tags$tag_examples[[1]], env = new.env(parent = globalenv(", "    topic = tools::file_path_sans_ext(topic$file_in), run_examples = e", "    run_dont_run = run_dont_run)")
#> highlight_examples(code, topic, env = env)
#> c("downlit::evaluate_and_highlight(code, fig_save = fig_save_topic, ", "    env = child_env(env))")
#> evaluate::evaluate(code, child_env(env), new_device = TRUE)
#> c("evaluate_call(expr, parsed$src[[i]], envir = envir, enclos = enclos, ", "    debug = debug, last = i == length(out), use_try = stop_on_error !=", "        2L, keep_warning = keep_warning, keep_message = keep_message, ", "    output_handler = output_handler, include_timing = include_timing)")
#> c("timing_fn(handle(ev <- withCallingHandlers(withVisible(eval(expr, ", "    envir, enclos)), warning = wHandler, error = eHandler, message = m")
#> c("handle(ev <- withCallingHandlers(withVisible(eval(expr, envir, ", "    enclos)), warning = wHandler, error = eHandler, message = mHandler")
#> try(f, silent = TRUE)
#> c("tryCatch(expr, error = function(e) {", "    call <- conditionCall(e)", "    if (!is.null(call)) {", "        if (identical(call[[1L]], quote(doTryCatch))) ", "            call <- sys.call(-4L)", "        dcall <- deparse(call)[1L]", "        prefix <- paste(\"Error in\", dcall, \": \")", "        LONG <- 75L", "        sm <- strsplit(conditionMessage(e), \"\\n\")[[1L]]", "        w <- 14L + nchar(dcall, type = \"w\") + nchar(sm[1L], type = \"w\"", "        if (is.na(w)) ", "            w <- 14L + nchar(dcall, type = \"b\") + nchar(sm[1L], ", 
#> "                type = \"b\")", "        if (w > LONG) ", "            prefix <- paste0(prefix, \"\\n  \")", "    }", "    else prefix <- \"Error : \"", "    msg <- paste0(prefix, conditionMessage(e), \"\\n\")", "    .Internal(seterrmessage(msg[1L]))", "    if (!silent && isTRUE(getOption(\"show.error.messages\"))) {", "        cat(msg, file = outFile)", "        .Internal(printDeferredWarnings())", "    }", "    invisible(structure(msg, class = \"try-error\", condition = e))", "})")
#> tryCatchList(expr, classes, parentenv, handlers)
#> tryCatchOne(expr, names, parentenv, handlers[[1L]])
#> doTryCatch(return(expr), name, parentenv, handler)
#> c("withCallingHandlers(withVisible(eval(expr, envir, enclos)), warning = ", "    error = eHandler, message = mHandler)")
#> withVisible(eval(expr, envir, enclos))
#> eval(expr, envir, enclos)
#> eval(expr, envir, enclos)
#> print(fm_CRSargs(crs2))
#> fm_CRSargs(crs2)
#> fm_not_for_PROJ6("fm_CRSargs")

#> Warning: Discarded ellps unknown in CRS definition: +proj=longlat +R=1 +no_defs +type=crs

#> Warning: Discarded datum unknown in CRS definition

#> [1] "+proj=longlat +R=1 +no_defs"