Create hexagon lattice points

Create hexagon lattice points within a boundary

Usage

fm_hexagon_lattice_orig(bnd, x_bin = 250, edge_len_n = 1)

Arguments

bnd

Boundary object

x_bin

Number of bins in x axis

edge_len_n

Number of edge length of mesh from the boundary to create hexagon mesh using x_bin

Value

A list with lattice points, edge length, and inner boundary

Author

Man Ho Suen M.H.Suen@sms.ed.ac.uk