Integration scheme for mesh triangle interiors

Integration scheme for mesh triangle interiors

Usage

fm_int_mesh_2d_core(mesh, tri_subset = NULL, nsub = NULL)

Arguments

mesh

Mesh on which to integrate

tri_subset

Optional triangle index vector for integration on a subset of the mesh triangles (Default NULL)

nsub

number of subdivision points along each triangle edge, giving (nsub + 1)^2 proto-integration points used to compute the vertex weights (default nsub=9, giving 100 integration points for each triangle)

Value

list with elements loc and weight with integration points for the mesh

Author

Finn Lindgren finn.lindgren@gmail.com

Examples

str(fm_int_mesh_2d_core(fmexample$mesh))
#> List of 2
#>  $ loc   : num [1:56300, 1:3] -0.717 -0.745 -0.774 -0.803 -0.832 ...
#>  $ weight: num [1:56300] 0.000733 0.000733 0.000733 0.000733 0.000733 ...