Make a 2D mesh object

Make a 2D mesh object

Usage

fm_mesh_2d(...)

fm_mesh_2d_inla(
  loc = NULL,
  loc.domain = NULL,
  offset = NULL,
  n = NULL,
  boundary = NULL,
  interior = NULL,
  max.edge = NULL,
  min.angle = NULL,
  cutoff = 1e-12,
  max.n.strict = NULL,
  max.n = NULL,
  plot.delay = NULL,
  crs = NULL,
  ...
)

Arguments

...

Currently passed on to fm_mesh_2d_inla

loc

Matrix of point locations to be used as initial triangulation nodes. Can alternatively be a SpatialPoints or SpatialPointsDataFrame object.

loc.domain

Matrix of point locations used to determine the domain extent. Can alternatively be a SpatialPoints or SpatialPointsDataFrame object.

offset

The automatic extension distance. One or two values, for an inner and an optional outer extension. If negative, interpreted as a factor relative to the approximate data diameter (default=-0.10???)

n

The number of initial nodes in the automatic extensions (default=16)

boundary

one or more (as list) of fm_segm() objects, or objects supported by fm_as_segm()

interior

one object supported by fm_as_segm()

max.edge

The largest allowed triangle edge length. One or two values.

min.angle

The smallest allowed triangle angle. One or two values. (Default=21)

cutoff

The minimum allowed distance between points. Point at most as far apart as this are replaced by a single vertex prior to the mesh refinement step.

max.n.strict

The maximum number of vertices allowed, overriding min.angle and max.edge (default=-1, meaning no limit). One or two values, where the second value gives the number of additional vertices allowed for the extension.

max.n

The maximum number of vertices allowed, overriding max.edge only (default=-1, meaning no limit). One or two values, where the second value gives the number of additional vertices allowed for the extension.

plot.delay

On Linux (and Mac if appropriate X11 libraries are installed), specifying a nonnegative numeric value activates a rudimentary plotting system in the underlying fmesher program, showing the triangulation algorithm at work, with waiting time factor plot.delay between each step.

On all systems, specifying any negative value activates displaying the result after each step of the multi-step domain extension algorithm.

crs

An optional fm_crs(), sf::crs or sp::CRS object

Value

An inla.mesh object.

Functions

fm_mesh_2d_inla(): Legacy method for INLA::inla.mesh.2d() Create a triangle mesh based on initial point locations, specified or automatic boundaries, and mesh quality parameters.

See also

Author

Finn Lindgren finn.lindgren@gmail.com