Make a tensor product function space

Tensor product function spaces. The interface and object storage model is experimental and may change.

Usage

fm_tensor(x, ...)

Arguments

x

list of function space objects, such as fm_mesh_2d().

...

Currently unused

Value

A fm_tensor or fm_tensor_list object. Elements of fm_tensor:

fun_spaces

fm_list of function space objects

manifold

character; manifold type summary. Regular subset of Rd "Rd", if all function spaces have type "R", torus connected "Td" if all function spaces have type "S", and otherwise "Md" In all cases, d is the sum of the manifold dimensions of the function spaces.

See also

Examples

m <- fm_tensor(list(
  space = fmexample$mesh,
  time = fm_mesh_1d(1:5)
))
m2 <- fm_as_tensor(m)
m3 <- fm_as_tensor_list(list(m, m))
c(fm_dof(m$fun_spaces$space) * fm_dof(m$fun_spaces$time), fm_dof(m))
#> [1] 1490 1490
str(fm_evaluator(m, loc = list(space = cbind(0, 0), time = 2.5)))
#> List of 1
#>  $ proj:List of 2
#>   ..$ A :Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. .. ..@ i       : int [1:6] 0 0 0 0 0 0
#>   .. .. ..@ p       : int [1:1491] 0 0 0 0 0 0 0 0 0 0 ...
#>   .. .. ..@ Dim     : int [1:2] 1 1490
#>   .. .. ..@ Dimnames:List of 2
#>   .. .. .. ..$ : NULL
#>   .. .. .. ..$ : NULL
#>   .. .. ..@ x       : num [1:6] 0.2394 0.2112 0.0494 0.2394 0.2112 ...
#>   .. .. ..@ factors : list()
#>   ..$ ok: logi TRUE
#>  - attr(*, "class")= chr "fm_evaluator"
str(fm_basis(m, loc = list(space = cbind(0, 0), time = 2.5)))
#> Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   ..@ i       : int [1:6] 0 0 0 0 0 0
#>   ..@ p       : int [1:1491] 0 0 0 0 0 0 0 0 0 0 ...
#>   ..@ Dim     : int [1:2] 1 1490
#>   ..@ Dimnames:List of 2
#>   .. ..$ : NULL
#>   .. ..$ : NULL
#>   ..@ x       : num [1:6] 0.2394 0.2112 0.0494 0.2394 0.2112 ...
#>   ..@ factors : list()
str(fm_fem(m))
#> List of 3
#>  $ cc:Formal class 'dgTMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:1490] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ j       : int [1:1490] 0 1 2 3 4 5 6 7 8 9 ...
#>   .. ..@ Dim     : int [1:2] 1490 1490
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : NULL
#>   .. ..@ x       : num [1:1490] 0.279 0.199 0.333 0.213 0.282 ...
#>   .. ..@ factors : list()
#>  $ g1:Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:12474] 0 1 26 244 256 298 0 1 2 256 ...
#>   .. ..@ p       : int [1:1491] 0 6 11 17 22 28 34 39 45 51 ...
#>   .. ..@ Dim     : int [1:2] 1490 1490
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : NULL
#>   .. ..@ x       : num [1:12474] 1.654 -0.332 -0.474 -0.188 -0.103 ...
#>   .. ..@ factors : list()
#>  $ g2:Formal class 'dgCMatrix' [package "Matrix"] with 6 slots
#>   .. ..@ i       : int [1:45672] 0 1 2 25 26 28 31 244 245 248 ...
#>   .. ..@ p       : int [1:1491] 0 21 36 56 73 93 111 128 148 167 ...
#>   .. ..@ Dim     : int [1:2] 1490 1490
#>   .. ..@ Dimnames:List of 2
#>   .. .. ..$ : NULL
#>   .. .. ..$ : NULL
#>   .. ..@ x       : num [1:45672] 11.457 -4.159 0.624 0.145 -4.716 ...
#>   .. ..@ factors : list()