fmesher: Triangle Meshes and Related Geometry Tools

Generate planar and spherical triangle meshes, compute finite element calculations for 1- and 2-dimensional flat and curved manifolds with associated basis function spaces, methods for lines and polygons, and transparent handling of coordinate reference systems and coordinate transformation, including 'sf' and 'sp' geometries. The core 'fmesher' library code was originally part of the 'INLA' package, and implements parts of "Triangulations and Applications" by Hjelle and Daehlen (2006) doi:10.1007/3-540-33261-8 .

See also

Author

Maintainer: Finn Lindgren finn.lindgren@gmail.com (ORCID) (Finn Lindgren wrote the main code) [copyright holder]

Other contributors:

Seaton Andy andy.e.seaton@gmail.com (Andy Seaton constributed features to the sf support) [contributor]

Suen Man Ho M.H.Suen@sms.ed.ac.uk (Man Ho Suen contributed features and code structure design for the integration methods) [contributor]

Fabian E. Bachl bachlfab@gmail.com (Fabian Bachl co-developed precursors of fm_pixels and fm_split_lines in inlabru) [contributor]