Diameter bound for a geometric object

Find an upper bound to the convex hull of a point set

Usage

fm_diameter(x, ...)

# S3 method for class 'matrix'
fm_diameter(x, manifold = NULL, ...)

# S3 method for class 'sf'
fm_diameter(x, ...)

# S3 method for class 'sfg'
fm_diameter(x, ...)

# S3 method for class 'sfc'
fm_diameter(x, ...)

# S3 method for class 'fm_lattice_2d'
fm_diameter(x, ...)

# S3 method for class 'fm_segm'
fm_diameter(x, ...)

# S3 method for class 'fm_mesh_2d'
fm_diameter(x, ...)

# S3 method for class 'fm_mesh_1d'
fm_diameter(x, ...)

# S3 method for class 'inla.mesh.1d'
fm_diameter(x, ...)

# S3 method for class 'inla.mesh.segment'
fm_diameter(x, ...)

# S3 method for class 'inla.mesh.lattice'
fm_diameter(x, ...)

# S3 method for class 'inla.mesh'
fm_diameter(x, ...)

Arguments

x

A point set as an \(n\times d\) matrix, or an fm_mesh_2d/1d/sf related object.

...

Additional parameters passed on to the submethods.

manifold

Character string specifying the manifold type. Default is to treat the point set with Euclidean \(R^d\) metrics. Use manifold="S2" for great circle distances on the unit sphere (this is set automatically for fm_fmesh_2d objects).

Value

A scalar, upper bound for the diameter of the convex hull of the point set.

Author

Finn Lindgren finn.lindgren@gmail.com

Examples

fm_diameter(matrix(c(0, 1, 1, 0, 0, 0, 1, 1), 4, 2))
#> [1] 1.414214