Non-convex hull computation

Legacy method for INLA::inla.nonconvex.hull()

Usage

fm_nonconvex_hull_inla(
  x,
  convex = -0.15,
  concave = convex,
  resolution = 40,
  eps = NULL,
  eps_rel = NULL,
  crs = NULL,
  ...
)

fm_nonconvex_hull_inla_basic(
  x,
  convex = -0.15,
  resolution = 40,
  eps = NULL,
  crs = NULL
)

Arguments

x

A spatial object

convex

numeric vector; How much to extend

concave

numeric vector; The minimum allowed reentrant curvature. Default equal to convex

resolution

The internal computation resolution. A warning will be issued when this needs to be increased for higher accuracy, with the required resolution stated.

eps, eps_rel

The polygonal curve simplification tolerances used for simplifying the resulting boundary curve. See fm_simplify_helper() for details.

crs

Options crs object for the resulting polygon

...

Unused.

Value

fm_nonconvex_hull_inla() returns an fm_segm/inla.mesh.segment

object, for compatibility with inla.nonconvex.hull().

Details

Requires splancs::nndistF()

Functions

fm_nonconvex_hull_inla_basic(): Special method for convex = 0.

INLA compatibility

For mesh and curve creation, the fm_rcdt_2d_inla(), fm_mesh_2d_inla(), and fm_nonconvex_hull_inla() methods will keep the interface syntax used by INLA::inla.mesh.create(), INLA::inla.mesh.2d(), and INLA::inla.nonconvex.hull() functions, respectively, whereas the fm_rcdt_2d(), fm_mesh_2d(), and fm_nonconvex_hull() interfaces may be different, and potentially change in the future.

See also

Examples

fm_nonconvex_hull_inla(cbind(0, 0), convex = 1)
#> fm_segm object:
#>   32 interior edges (1 group: 1)
#>   Bounding box = (-0.9992483, 0.9992483) x (-0.9992483, 0.9992483) x (0,0)