Unit test helpers

Local helper functions for package unit tests

local_testthat_assign(x, values, envir = parent.frame())

local_set_PROJ6_warnings(proj4 = FALSE, thin = TRUE, envir = parent.frame())

local_get_rgdal_options()

local_disable_PROJ6_warnings(envir = parent.frame())

local_fm_safe_inla(multicore = FALSE, quietly = TRUE, envir = parent.frame())

local_fm_testthat_setup(envir = parent.frame())

Arguments

x
character; Name of variable to assign to

values
the object to assign to x

envir
environment for exit handlers

proj4
logical; whether to show PROJ4 conversion warnings. Default FALSE

thin
logical; whether to show only a thinned version of rgdal PROJ6 warnings. Default TRUE

multicore
logical; if TRUE, multiple cores are allowed, and the INLA num.threads option is not checked or altered. Default: FALSE, multicore not allowed (used for examples and unit tests).

quietly
logical; if TRUE, prints diagnostic messages. A message is always printed if the INLA num.threads option is altered, regardless of the quietly argument. Default: TRUE.

Functions

local_testthat_assign: Assign local variable. Useful for easy cleanup of global workspace with withr::deferred_run() when running tests interactively.

local_set_PROJ6_warnings: Disable PROJ4/6 warnings. To be used within package tests. Restores state on exit.

local_get_rgdal_options: Return a list of the current rgdal warning options

local_disable_PROJ6_warnings: Disable rgdal PROJ4 conversion warnings and thin PROJ6 warnings.

local_fm_safe_inla: Tests should set num.threads = "1:1" to ensure within-system repeatability by calling local_fm_safe_inla(); see also fm_safe_inla()

local_fm_testthat_setup: Initialise environment for tests. Disables PROJ4/PROJ6 warnings, and assigns tolerance variables. To be called either at the top of a testfile, or inside tests. Does not call local_fm_safe_inla(), since that may invoke a skip and should be called inside each test that relies on INLA.